Following best practices for controller endpoint desiging
