+++ src/main/java/org/webshar/hrms/controller/OrganizationController.java
@@ -19,10 +19,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.webshar.hrms.model.db.Organization;
+import org.webshar.hrms.model.db.OrganizationLeave;
 import org.webshar.hrms.request.organization.OrganizationCreateRequest;
 import org.webshar.hrms.request.organization.OrganizationUpdateRequest;
+import org.webshar.hrms.request.organizationleave.OrganizationLeaveCreateRequest;
+import org.webshar.hrms.request.organizationleave.OrganizationLeaveUpdateRequest;
 import org.webshar.hrms.response.BatchResponse;
 import org.webshar.hrms.response.Response;
+import org.webshar.hrms.service.OrganizationLeaveService;
 import org.webshar.hrms.service.OrganizationService;
 import org.webshar.hrms.service.exception.EntityAlreadyExistsException;
 import org.webshar.hrms.service.exception.EntityNotFoundException;
@@ -35,6 +39,9 @@ public class OrganizationController
 
   @Autowired
   OrganizationService organizationService;
+
+  @Autowired
+  OrganizationLeaveService organizationLeaveService;
 
   private static final Logger LOGGER = LoggerFactory.getLogger(OrganizationController.class);
 
@@ -95,6 +102,71 @@ public class OrganizationController
     batchResponse.setEntities(Collections.singletonList(organizationsList));
     batchResponse.setTotalEntries(organizationsList.size());
     batchResponse.setMessage("Organizations fetched");
+    batchResponse.setStatus("OK");
+    return ResponseEntity.ok(batchResponse);
+  }
+
+  @GetMapping(value = "/services/api/web/hrms/organizations/{organization_id}/leaves")
+  public ResponseEntity<BatchResponse> getOrganizationLeavesByOrganizationId(
+      @PathVariable("organization_id") Long organizationId)
+      throws ServiceException
+  {
+    BatchResponse response = new BatchResponse();
+    List<OrganizationLeave> organizationLeaveList = organizationLeaveService
+        .getOrganizationLeaveByOrganizationId(organizationId);
+    response.setEntities(Collections.singletonList(organizationLeaveList));
+    response.setMessage("Organization leaves fetched");
+    response.setStatus("OK");
+    return ResponseEntity.ok(response);
+  }
+
+  @PostMapping(value = "/services/api/web/hrms/organizations/leaves")
+  public ResponseEntity<Response> createOrganizationLeave(
+      @Valid @RequestBody @NotNull OrganizationLeaveCreateRequest organizationLeaveCreateRequest)
+      throws ServiceException
+  {
+    Response response = new Response();
+    OrganizationLeave organizationLeave = organizationLeaveService
+        .createOrganizationLeave(organizationLeaveCreateRequest);
+    response.setEntity(organizationLeave);
+    response.setMessage("Organization leave added");
+    response.setStatus("OK");
+    return ResponseEntity.status(HttpStatus.CREATED).body(response);
+  }
+
+  @PatchMapping(value = "/services/api/web/hrms/organizations/{organization_id}/leaves")
+  public ResponseEntity<Response> updateOrganizationLeaveById(
+      @PathVariable("organization_id") Long organizationId,
+      @NotNull @Valid @RequestBody
+          OrganizationLeaveUpdateRequest organizationLeaveUpdateRequest) throws ServiceException
+  {
+    Response response = new Response();
+    OrganizationLeave organizationLeave = organizationLeaveService
+        .updateOrganizationLeave(organizationLeaveUpdateRequest);
+    response.setEntity(organizationLeave);
+    response.setMessage("Organization leave updated");
+    response.setStatus("OK");
+    return ResponseEntity.ok(response);
+  }
+
+  @DeleteMapping(value = "/services/api/web/hrms/organizations/{organization_id}/leaves")
+  public ResponseEntity<Void> deleteOrganizationLeaveById(
+      @PathVariable("organization_id") Long organizationId)
+      throws ServiceException
+  {
+    organizationLeaveService.deleteOrganizationLeaveById(organizationId);
+    return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+  }
+
+  @GetMapping(value = "/services/api/web/hrms/organizations/leaves")
+  public ResponseEntity<BatchResponse> getAllOrganizationLeaves()
+  {
+    BatchResponse batchResponse = new BatchResponse();
+    List<OrganizationLeave> organizationLeaveList;
+    organizationLeaveList = organizationLeaveService.getAllOrganizationLeaves();
+    batchResponse.setEntities(Collections.singletonList(organizationLeaveList));
+    batchResponse.setTotalEntries(organizationLeaveList.size());
+    batchResponse.setMessage("Organization leaves fetched");
     batchResponse.setStatus("OK");
     return ResponseEntity.ok(batchResponse);
   }
